docs: document profiler workflow and organize launcher
