+++ .gitignore
@@ -423,46 +423,6 @@ build/native/
 *.tlog
 *.lastbuildstate
 
-# Build outputs
-**/bin/
-**/obj/
-build/native/
-native/x64/
-native/Debug/
-native/Release/
-native/out/
-
-# Native compiler intermediates
-*.obj
-*.lib
-*.exp
-*.pdb
-*.ilk
-*.tlog
-*.lastbuildstate
-
-# Restored packages
-packages/
-
-# Local diagnostics and captured installations
-artifacts/
-
-# Visual Studio
-.vs/
-*.user
-*.suo
-*.VC.db
-*.VC.VC.opendb
-
-# Tests
-TestResults/
-
 # Local logs and temporary files
 *.log
-*.dmp
 *.tmp
-*.bak
-
-# Local README analysis
-_readme_context/
-readme-context.zip
